move lists and maps to utils.dart
